requirements.txt
Flask==3.1.0
pymongo==4.10.1
pydantic==2.12.5
python-dotenv==1.0.0
gunicorn==21.2.0
